backend/cloud-functions/review_harmful_content/requirements.txt
functions-framework==3.5.0
Flask==2.3.3
google-cloud-pubsub==2.18.0
google-cloud-firestore==2.11.0
